Extension: extract explainSelection command (oh-tab-b7qq)
